spostate funzioni di utilità da prenotazione.c a utile_prenotazione.c
Separata la logica di supporto relativa alle prenotazioni in un nuovo file `utile_prenotazione.c`
Migliorato l'information hiding e la modularità del codice
Aggiornati i riferimenti nei file header se necessario
