+++ prenotazione.h
@@ -1,7 +1,8 @@
-#ifndef PRENOTAZIONI_H
-#define PRENOTAZIONI_H
+#ifndef PRENOTAZIONE_H
+#define PRENOTAZIONE_H
 
 #include <stdbool.h>
+#define CELLE_GIORNALIERE 48  // 48 mezz'ore in un giorno
 
 /*
  Tipo astratto di dato: prenotazioni
@@ -35,97 +36,6 @@ ptr_prenotazione inizializza_prenotazioni();
 
 
 /*
- Funzione: carica_prenotazioni_da_file
- ------------------------------------
- Legge da file il contenuto delle prenotazioni per un veicolo,
- gestendo il reset automatico delle celle se è cambiato il giorno.
-
- Parametri:
-   p: puntatore alla struttura prenotazioni da aggiornare.
-   targa: stringa contenente la targa del veicolo, usata per costruire il nome file.
-
- Pre-condizione:
-   p deve essere un puntatore valido.
-   targa deve essere una stringa valida.
-
- Post-condizione:
-   La struttura prenotazioni è aggiornata con i dati letti dal file,
-   oppure azzerata se il file non esiste o il giorno è cambiato.
-   Le celle passate vengono bloccate.
-
- Ritorna:
-   1 se la lettura da file è avvenuta con successo,
-   0 se il file non esiste o il giorno è cambiato (prenotazioni ripristinate).
-*/
-int carica_prenotazioni_da_file(ptr_prenotazione p, const char *targa);
-
-
-/*
- Funzione: leggi_cella_da_orario
- -------------------------------
- Richiede all’utente di inserire un orario (HH MM),
- verifica la correttezza e lo converte in indice di slot.
-
- Parametri:
-   messaggio: stringa da visualizzare come prompt per l’inserimento.
-
- Pre-condizione:
-   Nessuna.
-
- Post-condizione:
-   Nessuna modifica.
-
- Ritorna:
-   Indice della cella corrispondente all’orario,
-   oppure -1 se l’orario inserito è non valido.
-*/
-int leggi_cella_da_orario(const char *messaggio);
-
-
-/*
- Funzione: veicolo_disponibile_oggi
- ----------------------------------
- Verifica se esiste almeno una cella libera oggi.
-
- Parametri:
-   p: puntatore alla struttura prenotazioni da controllare.
-
- Pre-condizione:
-   p deve essere un puntatore valido.
-
- Post-condizione:
-   Nessuna modifica.
-
- Ritorna:
-   true se almeno una cella è libera,
-   false altrimenti.
-*/
-bool veicolo_disponibile_oggi(ptr_prenotazione p);
-
-
-/*
- Funzione: salva_prenotazioni_su_file
- -----------------------------------
- Salva su file le prenotazioni aggiornate di un veicolo.
-
- Parametri:
-   p: puntatore alla struttura prenotazioni da salvare.
-   targa: stringa contenente la targa del veicolo, usata per costruire il nome file.
-
- Pre-condizione:
-   p deve essere un puntatore valido.
-   targa deve essere una stringa valida.
-
- Post-condizione:
-   Il file con le prenotazioni viene aggiornato.
-
- Ritorna:
-   Nessun valore.
-*/
-void salva_prenotazioni_su_file(ptr_prenotazione p, const char *targa);
-
-
-/*
  Funzione: prenota_intervallo
  ----------------------------
  Tenta di prenotare un intervallo di celle consecutive.
@@ -150,23 +60,70 @@ int prenota_intervallo(ptr_prenotazione p, int inizio_cella, int fine_cella);
 
 
 /*
- Funzione: libera_prenotazioni
- -----------------------------
- Libera la memoria allocata per la struttura prenotazioni.
+ Funzione: ottiene_cella
+ -----------------------
+ Restituisce lo stato di una specifica cella.
 
  Parametri:
-   p: puntatore alla struttura prenotazioni da liberare.
+   p: puntatore alla struttura prenotazioni.
+   indice: indice della cella da leggere.
+
+ Pre-condizione:
+   p deve essere un puntatore valido.
+   indice deve essere un indice valido.
+
+ Post-condizione:
+   Nessuna modifica alla struttura prenotazioni.
+
+ Ritorna:
+   1 se la cella è prenotata,
+   0 se la cella è libera,
+   -1 se p è NULL o indice non valido.
+*/
+int ottiene_cella (ptr_prenotazione p, int indice);
+
+
+/*
+ Funzione: imposta_cella
+ -----------------------
+ Imposta lo stato di una specifica cella.
+
+ Parametri:
+   p: puntatore alla struttura prenotazioni.
+   indice: indice della cella da modificare.
+   valore: valore da assegnare (0 o 1).
+
+ Pre-condizione:
+   p deve essere un puntatore valido.
+   indice deve essere un indice valido.
+
+ Post-condizione:
+   La cella specificata viene modificata con il valore dato.
+
+ Ritorna:
+   Nessun valore di ritorno.
+*/
+void imposta_cella(ptr_prenotazione p, int indice, int valore);
+
+
+/*
+ Funzione: azzera_celle
+ ----------------------
+ Azzera tutte le celle della struttura prenotazioni.
+
+ Parametri:
+   p: puntatore alla struttura prenotazioni.
 
  Pre-condizione:
    p deve essere un puntatore valido.
 
  Post-condizione:
-   La memoria della struttura è liberata.
+   Tutte le celle sono impostate a 0 (libere).
 
  Ritorna:
-   Nessun valore.
+   Nessun valore di ritorno.
 */
-void libera_prenotazioni(ptr_prenotazione p);
+void azzera_celle(ptr_prenotazione p);
 
 
 /*
@@ -191,28 +148,23 @@ void blocca_celle_passate(ptr_prenotazione p);
 
 
 /*
- Funzione: costo_noleggio
- -------------------------
- Calcola il costo totale del noleggio di un veicolo in base 
- all’intervallo orario specificato e ad un eventuale sconto.
+ Funzione: in_intervallo
+ -----------------------
+ Verifica se un indice rientra in un intervallo [inizio, fine).
 
  Parametri:
-   inizio_cella: indice della cella iniziale del noleggio (inclusivo).
-   fine_cella: indice della cella finale del noleggio (esclusivo).
-   sconto: valore intero. Se multiplo di 5, applica una tariffa scontata.
+   indice: indice della cella da verificare.
+   inizio: inizio dell'intervallo (inclusivo).
+   fine: fine dell'intervallo (esclusivo).
 
  Pre-condizione:
-   inizio_cella e fine_cella devono essere compresi tra 0 e 48,
-   con inizio_cella < fine_cella.
-   Il valore di sconto deve essere >= 0.
-
- Post-condizione:
-   Viene stampato il costo totale del noleggio.
-   Il valore di ritorno rappresenta il costo calcolato.
+   Tutti i parametri devono essere validi (0 <= inizio <= fine <= CELLE_GIORNALIERE).
 
  Ritorna:
-   Il costo del noleggio come valore float.
+   1 se indice è nell'intervallo [inizio, fine),
+   0 altrimenti.
 */
-float costo_noleggio(int inizio_cella, int fine_cella, int sconto);
+int in_intervallo(int indice, int inizio, int fine);
+
 
 #endif
